+++ package.json
@@ -9,6 +9,7 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "date-and-time": "^0.12.0"
+    "date-and-time": "^0.12.0",
+    "yargs": "^15.1.0"
   }
 }
